ルーティングとアプリ分けを変更
