+++ StageAPI/urls.py
@@ -1,0 +1,6 @@
+from rest_framework import routers
+from .views import StageViewSet
+
+
+router = routers.DefaultRouter()
+router.register(r"stages", StageViewSet)
