--- api/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-from .models import Stage
-
-
-@admin.register(Stage)
-class StageAdmin(admin.ModelAdmin):
-    pass
