--- api/urls.py
@@ -1,6 +1,0 @@
-from rest_framework import routers
-from .views import StageViewSet
-
-
-router = routers.DefaultRouter()
-router.register(r"stages", StageViewSet)
