+++ code_party_back/settings.py
@@ -36,7 +36,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "api",
+    "StageAPI",
     "testapp",
     "rest_framework",
 ]
